Expanding The World Map
I have expanded the world map. Not only have I made it bigger I have given it a structure that could work for the game and is scalable.

I have also added the tiles for the bosses and for the enemies that are specific for the bosses.

I fixed a couple of bugs that i came across. Not all bugs have been squashed but I think i have taken down some pretty noticable ones.

I have also made minor adjustments here and there to how the game outputs text. Eventually i want everything to look uniform so it have a level of neatness.