Adding Quests
I ahev added quests to the game. The map now has a tile to interact with the "quester" and i tile to find the item that the quester wants.

I am still in the process of expanding the world map. For now the world map  only has the first boss.