Story, Enemies, Weapons and other changes
The development of the story has begun. The story has been added through NPCs that the player can interact with. The new script, story.py stores all the NPC messages.

The enemy system has been changed. Enemies are now separated by tiers. This allows the game to become harder the more you progress through it.

With the addidtion of enemy tiers more weapons needed to be added to balance the game and add variety. This also means adjustments were made to the pricing of weapons and therefore the game econmy.

Other changes have been made to the presentation of the game. These changes allow for a better experience as there is greater uniformity.